Reorganize Graph Databases investigation directory
