+++ docs/investigation/graph-db/README.md
@@ -1,2 +1,6 @@
-# Neo4j Investigation
+# Graph database investigations
 
+The following directory contains different proof of concepts of using different graph databases for OCH.
+
+There are the following proof of concept projects:
+- [`neo4j-gogm`](./neo4j-gogm) - Using Neo4j database in OCH with GoGM library
