+++ docs/investigation/graph-db/neo4j-gogm/graphql/models_gen.go
