+++ docs/investigation/graph-db/neo4j-gogm/graphql/schema.graphql
