--- docs/investigation/graph-db/neo4j/ptr/ptr.go
@@ -1,17 +1,0 @@
-package ptr
-
-func String(in string) *string {
-	return &in
-}
-
-func Bool(in bool) *bool {
-	return &in
-}
-
-func Int32(in int32) *int32 {
-	return &in
-}
-
-func Int64(in int64) *int64 {
-	return &in
-}
